Organise misc Slippi connection components
