+++ src/renderer/containers/Settings/SlippiPage.tsx
@@ -1,10 +1,13 @@
 import * as React from "react";
 
-import { streamManager } from "@/lib/realtime";
-import { Dispatch, iRootState } from "@/store";
 import { ConnectionStatus } from "@vinceau/slp-realtime";
 import { useDispatch, useSelector } from "react-redux";
-import { ConnectionStatusCard, SlippiConnectionPlaceholder, statusToColor, statusToLabel } from "../Misc/ConnectionStatus";
+
+import { ConnectionStatusCard } from "@/components/ConnectionStatusCard";
+import { streamManager } from "@/lib/realtime";
+import { statusToColor, statusToLabel } from "@/lib/status";
+import { Dispatch, iRootState } from "@/store";
+import { SlippiConnectionPlaceholder } from "../SlippiConnectionPlaceholder";
 
 import slippiLogo from "@/styles/images/slippi.png";
